pyproject.toml
[tool.poetry]
name = "mgyminer"
version = "0.0.1"
description = "Platform to explore the MGnify protein database"
authors = ["Felix Langer <felix@ebi.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp-client-cache = "^0.7.3"
PyYAML = "^6.0"
mysql-connector-python = "^8.0.30"
numpy = "^1.23.2"
pandas = "^1.4.3"
biopython = "^1.79"
diskcache = "5.4.0"
plotly = "^5.10.0"
dash = "^2.6.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
pytest = "^7.1.2"

[tool.poetry.scripts]
MGnifyMiner = "mgyminer.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
